Separe actions and reducers albums and songs
