+++ src/App/Pages/AlbumDetail.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { connect } from 'react-redux';
 import SongsList from '../Model/SongsList';
 
-const AlbumDetail = ({server, history, location, match}) => {
-  if (server.songsLoading) {
+const AlbumDetail = ({songs, history, location, match}) => {
+  if (songs.isLoading) {
     return <p>Cargando...</p>
-  } else if (server.songsError) {
+  } else if (songs.hasError) {
     return <p>Hubo un error al obtener los datos :(</p>
   } else {
-    return  <SongsList songs={server.songs} 
+    return  <SongsList songs={songs.list} 
                        history={history}  
                        location={location} 
                        match={match} />
