+++ src/App/Pages/Albums.js
@@ -2,18 +2,16 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import AlbumList from '../Model/AlbumList';
 
-import { getAlbums } from '../actions/server';
-
 class Albums extends Component {
     renderAlbums() {
-        const { server } = this.props;
+        const { albums } = this.props;
     
-        if (server.albumsLoading) {
+        if (albums.isLoading) {
           return <p>Cargando...</p>
-        } else if (server.albumsError) {
+        } else if (albums.hasError) {
           return <p>Hubo un error al obtener los datos :(</p>
         } else {
-          return  <AlbumList albums={server.albums} 
+          return  <AlbumList albums={albums.list} 
                     history={this.props.history}  
                     location={this.props.location} 
                     match={this.props.match} />
