+++ src/App/store.js
@@ -4,9 +4,10 @@ import promise from 'redux-promise-middleware';
 // Reducers
 //import todos from './reducers/todos';
 import user from './reducers/user';
-import server from './reducers/server';
+import albums from './reducers/albums';
+import songs from './reducers/songs';
 
 export default createStore(
-    combineReducers({ user, server }), 
+    combineReducers({ user, albums, songs }), 
     applyMiddleware(promise)
 );
